+++ bidsme/Modules/MRI/jsonNIFTI.py
@@ -24,7 +24,6 @@
 
 import os
 import logging
-import shutil
 import json
 from datetime import datetime
 
@@ -113,14 +112,10 @@ class jsonNIFTI(MRI):
         return None
 
     def dump(self):
-        if self._HEADER_CACHE is not None:
-            return str(self._HEADER_CACHE)
-        elif len(self.files) > 0:
+        if self._HEADER_CACHE is None:
             self.loadFile(0)
-            return str(self._HEADER_CACHE)
-        else:
-            logger.error("No defined files")
-            return "No defined files"
+        res = self._HEADER_CACHE
+        return res
 
     def _getField(self, field: list):
         res = None
@@ -151,16 +146,6 @@ class jsonNIFTI(MRI):
     def clearCache(self) -> None:
         self._HEADER_CACHE = None
         self._FILE_CACHE = ""
-
-    def copyRawFile(self, destination: str) -> str:
-        if os.path.isfile(os.path.join(destination,
-                                       self.currentFile(True))):
-            logger.warning("{}: File {} exists at destination"
-                           .format(self.recIdentity(),
-                                   self.currentFile(True)))
-        shutil.copy2(self.currentFile(), destination)
-        shutil.copy2(self._header_file, destination)
-        return os.path.join(destination, self.currentFile(True))
 
     def _getSubId(self) -> str:
         return self.getField("PatientID", "")