+++ bidsme/Modules/MRS/jsonNIFTI.py
@@ -24,7 +24,6 @@
 
 import os
 import logging
-import shutil
 import json
 from datetime import datetime
 from copy import copy
@@ -120,14 +119,9 @@ class jsonNIFTI(MRS):
         return None
 
     def dump(self):
-        if self._HEADER_CACHE is not None:
-            return str(self._HEADER_CACHE)
-        elif len(self.files) > 0:
+        if self._HEADER_CACHE is None:
             self.loadFile(0)
-            return str(self._HEADER_CACHE)
-        else:
-            logger.error("No defined files")
-            return "No defined files"
+        return self._HEADER_CACHE
 
     def _getField(self, field: list):
         res = None
@@ -158,16 +152,6 @@ class jsonNIFTI(MRS):
     def clearCache(self) -> None:
         self._HEADER_CACHE = None
         self._FILE_CACHE = ""
-
-    def copyRawFile(self, destination: str) -> str:
-        if os.path.isfile(os.path.join(destination,
-                                       self.currentFile(True))):
-            logger.warning("{}: File {} exists at destination"
-                           .format(self.recIdentity(),
-                                   self.currentFile(True)))
-        shutil.copy2(self.currentFile(), destination)
-        shutil.copy2(self._header_file, destination)
-        return os.path.join(destination, self.currentFile(True))
 
     def _getSubId(self) -> str:
         return self.getField("PatientID", "")