+++ bidsme/gui/__init__.py
@@ -1,0 +1,3 @@
+from .MetaExplorer import MetaExplorer, cMetaExplorer
+
+__all__ = ["MetaExplorer", "cMetaExplorer"]